+++ src/main/java/org/example/payment/WechatPay.java
@@ -1,8 +1,10 @@
 package org.example.payment;
 
+import java.math.BigDecimal;
+
 public class WechatPay implements Payment {
     @Override
-    public boolean pay(float totalMoney) {
+    public boolean pay(BigDecimal totalMoney) {
         System.out.println("You should pay: " + totalMoney + " with Wechat");
         return true;
     }
